going_modular/requirements.txt
# Requirements to be downloaded
ipython==8.31.0
Pillow==11.1.0
requests==2.32.3
torch==2.5.1+cu124
torchvision==0.20.1+cu124
torchaudio==2.5.1
torchinfo==1.8.0
torchmetrics==1.6.1
tqdm==4.67.1
